@@ -1,6 +1,6 @@
 {
   "name": "pyodide",
-  "version": "0.29.0-dev.0",
+  "version": "0.30.0-dev.0",
   "description": "The Pyodide JavaScript package",
   "keywords": [
     "python",
@@ -15,27 +15,29 @@
     "url": "https://github.com/pyodide/pyodide/issues"
   },
   "license": "MPL-2.0",
+  "dependencies": {
+    "@types/emscripten": "^1.41.4",
+    "ws": "^8.5.0"
+  },
   "devDependencies": {
+    "@biomejs/biome": "2.1.1",
+    "@playwright/test": "^1.55.1",
     "@types/assert": "^1.5.6",
     "@types/expect": "^24.3.0",
-    "@types/mocha": "^9.1.0",
     "@types/node": "^20.8.4",
     "@types/ws": "^8.5.3",
-    "chai": "^4.3.6",
-    "chai-as-promised": "^7.1.1",
     "cross-env": "^7.0.3",
     "dts-bundle-generator": "^8.1.1",
     "esbuild": "^0.25.0",
     "express": "^4.17.3",
-    "mocha": "^9.0.2",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
+    "playwright": "^1.55.1",
     "prettier": "^2.2.1",
-    "sinon": "^18.0.0",
-    "ts-mocha": "^9.0.2",
     "tsd": "^0.24.1",
-    "typedoc": "^0.27.6",
-    "typescript": "5.7",
+    "tsx": "^4.20.5",
+    "typedoc": "^0.27.9",
+    "typescript": "5.8",
     "wabt": "^1.0.32"
   },
   "main": "pyodide.js",
@@ -66,7 +68,8 @@
     "pyodide.d.ts",
     "ffi.d.ts",
     "pyodide-lock.json",
-    "console.html"
+    "console.html",
+    "console-v2.html"
   ],
   "browser": {
     "child_process": false,
@@ -82,24 +85,12 @@
     "build-inner": "node esbuild.config.inner.mjs",
     "build": "tsc --noEmit && node esbuild.config.outer.mjs",
     "test": "npm-run-all test:*",
-    "test:unit": "cross-env TEST_NODE=1 ts-mocha --node-option=experimental-loader=./test/loader.mjs --node-option=experimental-wasm-stack-switching -p tsconfig.test.json \"test/unit/**\"",
-    "test:node": "cross-env TEST_NODE=1 mocha test/integration/**/*.test.js",
-    "test:browser": "mocha test/integration/**/*.test.js",
+    "test:unit": "cross-env TEST_NODE=1 node --import tsx --experimental-wasm-stack-switching --test 'test/unit/**/*.test.*'",
+    "test:node": "cross-env TEST_NODE=1 npx playwright test",
+    "test:browser": "npx playwright test",
     "tsc": "tsc --noEmit",
     "coverage": "cross-env TEST_NODE=1 npm-run-all coverage:*",
     "coverage:build": "nyc npm run test:node"
-  },
-  "mocha": {
-    "bail": false,
-    "timeout": 30000,
-    "full-trace": true,
-    "inline-diffs": true,
-    "check-leaks": false,
-    "global": [
-      "pyodide",
-      "page",
-      "chai"
-    ]
   },
   "nyc": {
     "reporter": [
@@ -126,9 +117,6 @@
         "DOM"
       ]
     }
-  },
-  "dependencies": {
-    "ws": "^8.5.0"
   },
   "types": "./pyodide.d.ts",
   "engines": {
