@@ -1,29 +1,26 @@
-const chai = require("chai");
+import { expect, test } from "./fixture";
 
-describe("Pyodide", () => {
-  it("runPython", async () => {
+test.describe("Pyodide", () => {
+  test("runPython", async ({ page }) => {
     const factory = async () => {
       return pyodide.runPython("1+1");
     };
-    const result = await chai.assert.isFulfilled(page.evaluate(factory));
-    chai.assert.equal(result, 2);
+    const result = await page.evaluate(factory);
+    expect(result).toStrictEqual(2);
   });
-  describe("micropip", () => {
-    before(async () => {
+  test.describe("micropip", () => {
+    test.beforeEach(async ({ page }) => {
       const factory = async () => {
         return pyodide.loadPackage(["micropip"]);
       };
-      const installedPackages = await chai.assert.isFulfilled(
-        page.evaluate(factory),
-      );
-      chai.assert.isNotEmpty(installedPackages);
-      chai.assert.include(
-        installedPackages.map((pkg) => pkg.name),
-        "micropip",
+      const installedPackages = await page.evaluate(factory);
+      expect(installedPackages.length).toBeGreaterThan(0);
+      expect(installedPackages.some((pkg) => pkg.name === "micropip")).toBe(
+        true,
       );
     });
 
-    it("install", async () => {
+    test("install", async ({ page }) => {
       const factory = async () => {
         await pyodide.runPythonAsync(
           'import micropip; await micropip.install("snowballstemmer")',
@@ -33,8 +30,8 @@ describe("Pyodide", () => {
           len(snowballstemmer.stemmer('english').stemWords(['A', 'node', 'test']))
         `);
       };
-      const result = await chai.assert.isFulfilled(page.evaluate(factory));
-      chai.assert.equal(result, 3);
+      const result = await page.evaluate(factory);
+      expect(result).toStrictEqual(3);
     });
   });
 });
