@@ -16,9 +16,7 @@ DIST_PATH = PYODIDE_ROOT / "dist"
 
 sys.path.append(str(PYODIDE_ROOT / "src" / "py"))
 
-# importing this fixture has a side effect of making the safari webdriver reused during the session
 from pytest_pyodide import get_global_config
-from pytest_pyodide.runner import use_global_safari_service  # noqa: F401
 from pytest_pyodide.utils import package_is_built as _package_is_built
 
 os.environ["IN_PYTEST"] = "1"
@@ -64,7 +62,7 @@ def set_configs():
         pyodide._api.pyodide_ffi.register_js_module;
         pyodide._api.pyodide_ffi.unregister_js_module;
         pyodide.pyimport("pyodide.ffi.wrappers").destroy();
-        pyodide.pyimport("pyodide.http").destroy();
+        pyodide.pyimport("pyodide.http.pyxhr").destroy();
         pyodide.pyimport("pyodide_js._api");
     """)
 
@@ -189,10 +187,25 @@ def pytest_collection_modifyitems(config, items):
         cache = config.cache
         prev_test_result = cache.get("cache/lasttestresult", {})
 
+    # Skip long_running tests unless in CI environment
+    is_ci = os.environ.get("CI", "").lower() in ("true", "1", "yes")
+
     for item in items:
         if prev_test_result.get(item.nodeid) in ("passed", "warnings", "skip_passed"):
             item.add_marker(pytest.mark.skip(reason="previously passed"))
             continue
+
+        # Skip long_running tests unless explicitly running them or in CI
+        if item.get_closest_marker("long_running") and not is_ci:
+            # Check if user explicitly wants to run long_running tests
+            markexpr = config.getoption("-m", default="")
+            if "long_running" not in markexpr:
+                item.add_marker(
+                    pytest.mark.skip(
+                        reason="long_running test skipped (use '-m long_running' to run or set CI=1)"
+                    )
+                )
+                continue
 
         maybe_skip_test(item, delayed=True)
 
