@@ -1,7 +1,7 @@
 from collections.abc import Iterator
 
 from sphinx_js import ir
-from sphinx_js.ir import Class
+from sphinx_js.ir import Class, TypeXRefInternal
 from sphinx_js.typedoc import Analyzer as TsAnalyzer
 
 __all__ = ["ts_xref_formatter", "patch_sphinx_js"]
@@ -16,7 +16,11 @@ def ts_xref_formatter(_config, xref):
     from sphinx_pyodide.mdn_xrefs import JSDATA
 
     name = xref.name
-    if name == "PyodideAPI":
+    if name == "Lockfile":
+        name = "~pyodide.Lockfile"
+    if name == "TypedArray":
+        name = "~pyodide.ffi.TypedArray"
+    if name in ("PyodideAPI", "PyodideAPI_"):
         return ":ref:`PyodideAPI <js-api-pyodide>`"
     if name in JSDATA:
         return f":js:data:`{name}`"
@@ -24,6 +28,8 @@ def ts_xref_formatter(_config, xref):
         return f":js:class:`~pyodide.ffi.{name}`"
     if name in ["ConcatArray", "IterableIterator", "unknown", "U"]:
         return f"``{name}``"
+    if isinstance(xref, TypeXRefInternal):
+        return f":js:{xref.kind}:`{name}`"
     return f":js:class:`{name}`"
 
 
@@ -67,7 +73,7 @@ def _get_toplevel_objects(
     )
     PYPROXY_METHODS.update(methodPairs)
     for obj in ir_objects:
-        if obj.name == "PyodideAPI":
+        if obj.name == "PyodideAPI_":
             for member in obj.members:
                 member.documentation_root = True
             yield from _get_toplevel_objects(self, obj.members)
@@ -76,6 +82,8 @@ def _get_toplevel_objects(
             continue
         if doclet_is_private(obj):
             continue
+        if has_tag(obj, "hidetype"):
+            obj.type = None
         mod = get_obj_mod(obj)
         set_kind(obj)
         if obj.deppath == "./core/pyproxy" and isinstance(obj, Class):
@@ -118,8 +126,11 @@ def get_obj_mod(doclet: ir.TopLevel) -> str:
     filename = key[0]
     doclet.name = doclet.name.rpartition(".")[2]
 
+    if kind := doclet.block_tags.get("docgroup"):
+        return kind[0][0].text
+
     if filename == "pyodide.":
-        return "globalThis"
+        return "exports"
 
     if filename == "canvas.":
         return "pyodide.canvas"
