@@ -1,14 +1,6 @@
 import ErrorStackParser from "./vendor/stackframe/error-stack-parser";
-import {
-  IN_NODE,
-  IN_NODE_ESM,
-  IN_BROWSER_MAIN_THREAD,
-  IN_BROWSER_WEB_WORKER,
-  IN_NODE_COMMONJS,
-  IN_SHELL,
-} from "./environments";
+import { RUNTIME_ENV } from "./environments";
 import { Lockfile } from "./types";
-
 let nodeUrlMod: typeof import("node:url");
 let nodePath: typeof import("node:path");
 let nodeVmMod: typeof import("node:vm");
@@ -25,6 +17,7 @@ declare var globalThis: {
   importScripts: (url: string) => void;
   document?: typeof document;
   fetch?: typeof fetch;
+  location?: URL;
 };
 
 /**
@@ -33,7 +26,7 @@ declare var globalThis: {
  * @private
  */
 export async function initNodeModules() {
-  if (!IN_NODE) {
+  if (!RUNTIME_ENV.IN_NODE) {
     return;
   }
   // @ts-ignore
@@ -74,6 +67,10 @@ export async function initNodeModules() {
   };
 }
 
+export function isAbsolute(path: string): boolean {
+  return path.includes("://") || path.startsWith("/");
+}
+
 function node_resolvePath(path: string, base?: string): string {
   return nodePath.resolve(base || ".", path);
 }
@@ -87,9 +84,9 @@ function browser_resolvePath(path: string, base?: string): string {
 }
 
 export let resolvePath: (rest: string, base?: string) => string;
-if (IN_NODE) {
+if (RUNTIME_ENV.IN_NODE) {
   resolvePath = node_resolvePath;
-} else if (IN_SHELL) {
+} else if (RUNTIME_ENV.IN_SHELL) {
   resolvePath = (x) => x;
 } else {
   resolvePath = browser_resolvePath;
@@ -102,7 +99,7 @@ if (IN_NODE) {
  */
 export let pathSep: string;
 
-if (!IN_NODE) {
+if (!RUNTIME_ENV.IN_NODE) {
   pathSep = "/";
 }
 
@@ -187,9 +184,9 @@ export let getBinaryResponse: (
 ) =>
   | { response: Promise<Response>; binary?: undefined }
   | { response?: undefined; binary: Promise<Uint8Array> };
-if (IN_NODE) {
+if (RUNTIME_ENV.IN_NODE) {
   getBinaryResponse = node_getBinaryResponse;
-} else if (IN_SHELL) {
+} else if (RUNTIME_ENV.IN_SHELL) {
   getBinaryResponse = shell_getBinaryResponse;
 } else {
   getBinaryResponse = browser_getBinaryResponse;
@@ -217,10 +214,10 @@ export async function loadBinaryFile(
  */
 export let loadScript: (url: string) => Promise<void>;
 
-if (IN_BROWSER_MAIN_THREAD) {
+if (RUNTIME_ENV.IN_BROWSER_MAIN_THREAD) {
   // browser
   loadScript = async (url) => await import(/* webpackIgnore: true */ url);
-} else if (IN_BROWSER_WEB_WORKER) {
+} else if (RUNTIME_ENV.IN_BROWSER_WEB_WORKER) {
   // webworker
   loadScript = async (url) => {
     try {
@@ -235,9 +232,9 @@ if (IN_BROWSER_MAIN_THREAD) {
       }
     }
   };
-} else if (IN_NODE) {
+} else if (RUNTIME_ENV.IN_NODE) {
   loadScript = nodeLoadScript;
-} else if (IN_SHELL) {
+} else if (RUNTIME_ENV.IN_SHELL) {
   loadScript = load;
 } else {
   throw new Error("Cannot determine runtime environment");
@@ -264,13 +261,13 @@ async function nodeLoadScript(url: string) {
 }
 
 export async function loadLockFile(lockFileURL: string): Promise<Lockfile> {
-  if (IN_NODE) {
+  if (RUNTIME_ENV.IN_NODE) {
     await initNodeModules();
     const package_string = await nodeFsPromisesMod.readFile(lockFileURL, {
       encoding: "utf8",
     });
     return JSON.parse(package_string);
-  } else if (IN_SHELL) {
+  } else if (RUNTIME_ENV.IN_SHELL) {
     const package_string = read(lockFileURL);
     return JSON.parse(package_string);
   } else {
@@ -284,7 +281,7 @@ export async function loadLockFile(lockFileURL: string): Promise<Lockfile> {
  * This is used to guess the indexURL when it is not provided.
  */
 export async function calculateDirname(): Promise<string> {
-  if (IN_NODE_COMMONJS) {
+  if (RUNTIME_ENV.IN_NODE_COMMONJS) {
     return __dirname;
   }
 
@@ -296,11 +293,11 @@ export async function calculateDirname(): Promise<string> {
   }
   let fileName = ErrorStackParser.parse(err)[0].fileName!;
 
-  if (IN_NODE && !fileName.startsWith("file://")) {
+  if (RUNTIME_ENV.IN_NODE && !fileName.startsWith("file://")) {
     fileName = `file://${fileName}`; // Error stack filenames are not starting with `file://` in `Bun`
   }
 
-  if (IN_NODE_ESM) {
+  if (RUNTIME_ENV.IN_NODE_ESM) {
     const nodePath = await import("node:path");
     const nodeUrl = await import("node:url");
 
@@ -312,7 +309,7 @@ export async function calculateDirname(): Promise<string> {
   const indexOfLastSlash = fileName.lastIndexOf(pathSep);
   if (indexOfLastSlash === -1) {
     throw new Error(
-      "Could not extract indexURL path from pyodide module location",
+      "Could not extract indexURL path from pyodide module location. Please pass the indexURL explicitly to loadPyodide.",
     );
   }
   return fileName.slice(0, indexOfLastSlash);
@@ -322,8 +319,11 @@ export async function calculateDirname(): Promise<string> {
  * Ensure that the directory exists before trying to download files into it (Node.js only).
  * @param dir The directory to ensure exists
  */
-export async function ensureDirNode(dir: string) {
-  if (!IN_NODE) {
+export async function ensureDirNode(dir?: string) {
+  if (!RUNTIME_ENV.IN_NODE) {
+    return;
+  }
+  if (!dir) {
     return;
   }
 
@@ -336,4 +336,23 @@ export async function ensureDirNode(dir: string) {
       recursive: true,
     });
   }
+}
+
+/**
+ * Calculates the install base url for the package manager.
+ * exported for testing
+ * @param lockFileURL
+ * @returns the install base url
+ * @private
+ */
+export function calculateInstallBaseUrl(lockFileURL: string) {
+  // 1. If the lockfile URL includes a path with slash (file url in Node.js or http url in browser), use the directory of the lockfile URL
+  // 2. Otherwise, fallback to the current location
+  //    2.1. In the browser, use `location` to get the current location
+  //    2.2. In Node.js just use the pwd
+  return (
+    lockFileURL.substring(0, lockFileURL.lastIndexOf("/") + 1) ||
+    globalThis.location?.toString() ||
+    "."
+  );
 }
