@@ -1,10 +1,11 @@
 /** @private */
 
-import { ConfigType } from "./pyodide";
+import { PyodideConfigWithDefaults } from "./pyodide";
 import { initializeNativeFS } from "./nativefs";
 import { loadBinaryFile, getBinaryResponse } from "./compat";
-import { API, PreRunFunc, type Module } from "./types";
+import { API, PreRunFunc, type PyodideModule, type FSType } from "./types";
 import { getSentinelImport } from "generated/sentinel";
+import { RUNTIME_ENV } from "./environments";
 
 /**
  * @private
@@ -40,7 +41,10 @@ export interface EmscriptenSettings {
  *
  * @private
  */
-export function createSettings(config: ConfigType): EmscriptenSettings {
+export function createSettings(
+  config: PyodideConfigWithDefaults,
+): EmscriptenSettings {
+  const API = { config, runtimeEnv: RUNTIME_ENV } as API;
   const settings: EmscriptenSettings = {
     noImageDecoding: true,
     noAudioDecoding: true,
@@ -53,7 +57,7 @@ export function createSettings(config: ConfigType): EmscriptenSettings {
     },
     thisProgram: config._sysExecutable,
     arguments: config.args,
-    API: { config } as API,
+    API,
     // Emscripten calls locateFile exactly one time with argument
     // pyodide.asm.wasm to get the URL it should download it from.
     //
@@ -105,7 +109,7 @@ function setEnvironment(env: { [key: string]: string }): PreRunFunc {
  * @param mounts The list of paths to mount.
  */
 function callFsInitHook(
-  fsInit: undefined | ((fs: typeof FS, info: { sitePackages: string }) => void),
+  fsInit: undefined | ((fs: FSType, info: { sitePackages: string }) => void),
 ): PreRunFunc[] {
   if (!fsInit) {
     return [];
@@ -122,7 +126,7 @@ function callFsInitHook(
   ];
 }
 
-function computeVersionTuple(Module: Module): [number, number, number] {
+function computeVersionTuple(Module: PyodideModule): [number, number, number] {
   const versionInt = Module.HEAPU32[Module._Py_Version >>> 2];
   const major = (versionInt >>> 24) & 0xff;
   const minor = (versionInt >>> 16) & 0xff;
@@ -144,7 +148,7 @@ function computeVersionTuple(Module: Module): [number, number, number] {
  */
 function installStdlib(stdlibURL: string): PreRunFunc {
   const stdlibPromise: Promise<Uint8Array> = loadBinaryFile(stdlibURL);
-  return async (Module: Module) => {
+  return async (Module: PyodideModule) => {
     Module.API.pyVersionTuple = computeVersionTuple(Module);
     const [pymajor, pyminor] = Module.API.pyVersionTuple;
     Module.FS.mkdirTree("/lib");
@@ -168,7 +172,9 @@ function installStdlib(stdlibURL: string): PreRunFunc {
  * Initialize the virtual file system, before loading Python interpreter.
  * @private
  */
-function getFileSystemInitializationFuncs(config: ConfigType): PreRunFunc[] {
+function getFileSystemInitializationFuncs(
+  config: PyodideConfigWithDefaults,
+): PreRunFunc[] {
   let stdLibURL;
   if (config.stdLibURL != undefined) {
     stdLibURL = config.stdLibURL;
