@@ -1,20 +1,19 @@
-import {
-  IN_BROWSER_MAIN_THREAD,
-  IN_NODE,
-  IN_DENO,
-  IN_SAFARI,
-} from "./environments";
+import { RUNTIME_ENV } from "./environments";
 
 const scheduleCallbackImmediateMessagePrefix =
   "sched$" + Math.random().toString(36).slice(2) + "$";
+
 const tasks: Record<number, () => void> = {};
 let nextTaskHandle = 0;
+
+let sharedChannel: MessageChannel | null = null;
+const taskQueue: number[] = [];
 
 /**
  * Setup global message event listener to handle immediate callbacks
  */
 function installPostMessageHandler() {
-  if (!IN_BROWSER_MAIN_THREAD) {
+  if (!RUNTIME_ENV.IN_BROWSER_MAIN_THREAD) {
     return;
   }
 
@@ -44,6 +43,34 @@ function installPostMessageHandler() {
 
 installPostMessageHandler();
 
+function ensureSharedChannel() {
+  if (sharedChannel) return;
+  if (RUNTIME_ENV.IN_SAFARI || RUNTIME_ENV.IN_DENO || RUNTIME_ENV.IN_NODE)
+    return;
+  if (typeof globalThis.MessageChannel !== "function") return;
+
+  sharedChannel = new MessageChannel();
+
+  sharedChannel.port1.onmessage = () => {
+    // Process tasks that were queued when this message arrived
+    const count = taskQueue.length;
+    for (let i = 0; i < count; i++) {
+      const handle = taskQueue.shift()!;
+      const task = tasks[handle];
+      if (!task) continue;
+      try {
+        task();
+      } finally {
+        delete tasks[handle];
+      }
+    }
+  };
+  // Redundant per spec (onmessage auto-starts the port); kept for clarity and potential edge runtimes.
+  sharedChannel.port1.start?.();
+}
+
+ensureSharedChannel();
+
 /**
  * Implementation of zero-delay scheduler for immediate callbacks
  * Try our best to use the fastest method available, based on the current environment.
@@ -60,34 +87,24 @@ installPostMessageHandler();
  *   - Ref: https://github.com/YuzuJS/setImmediate/issues/80
  */
 function scheduleCallbackImmediate(callback: () => void) {
-  if (IN_NODE) {
-    // node has setImmediate, let's use it
+  if (RUNTIME_ENV.IN_NODE) {
     setImmediate(callback);
+  } else if (sharedChannel) {
+    const handle = nextTaskHandle++;
+    tasks[handle] = callback;
+    taskQueue.push(handle);
+    sharedChannel.port2.postMessage(0);
   } else if (
-    !IN_SAFARI &&
-    !IN_DENO &&
-    typeof globalThis.MessageChannel === "function"
-  ) {
-    const channel = new MessageChannel();
-    channel.port1.onmessage = () => {
-      channel.port1.onmessage = null;
-      channel.port1.close();
-      channel.port2.close();
-      callback();
-    };
-    channel.port2.postMessage("");
-  } else if (
-    IN_BROWSER_MAIN_THREAD &&
+    RUNTIME_ENV.IN_BROWSER_MAIN_THREAD &&
     typeof globalThis.postMessage === "function"
   ) {
-    tasks[nextTaskHandle] = callback;
+    const handle = nextTaskHandle++;
+    tasks[handle] = callback;
     globalThis.postMessage(
-      scheduleCallbackImmediateMessagePrefix + nextTaskHandle,
+      scheduleCallbackImmediateMessagePrefix + handle,
       "*",
     );
-    nextTaskHandle++;
   } else {
-    // fallback to setTimeout if nothing else is available
     setTimeout(callback, 0);
   }
 }
@@ -96,6 +113,7 @@ function scheduleCallbackImmediate(callback: () => void) {
  * Schedule a callback. Supports both immediate and delayed callbacks.
  * @param callback The callback to be scheduled
  * @param timeout The delay in milliseconds before the callback is called
+ * @hidden
  */
 export function scheduleCallback(callback: () => void, timeout: number = 0) {
   if (timeout <= 2) {
