@@ -1,9 +1,9 @@
-import { IN_NODE, IN_SHELL } from "./environments.js";
+import { RUNTIME_ENV } from "./environments.js";
 import "./constants";
 
 import type { FSStream, FSStreamOpsGen } from "./types";
-const fs: any = IN_NODE ? require("node:fs") : undefined;
-const tty: any = IN_NODE ? require("node:tty") : undefined;
+const fs: any = RUNTIME_ENV.IN_NODE ? require("node:fs") : undefined;
+const tty: any = RUNTIME_ENV.IN_NODE ? require("node:tty") : undefined;
 
 function nodeFsync(fd: number): void {
   try {
@@ -12,10 +12,16 @@ function nodeFsync(fd: number): void {
     if (e?.code === "EINVAL") {
       return;
     }
-    // On mac, calling fsync on stdout/stderr when not isatty returns ENOTSUP
-    if (e?.code === "ENOTSUP" && (fd === 1 || fd === 2)) {
+    // On Mac, calling fsync when not isatty returns ENOTSUP
+    // On Windows, stdin/stdout/stderr may be closed, returning EBADF or EPERM
+    const isStdStream = fd === 0 || fd === 1 || fd === 2;
+    if (
+      isStdStream &&
+      (e?.code === "ENOTSUP" || e?.code === "EBADF" || e?.code === "EPERM")
+    ) {
       return;
     }
+
     throw e;
   }
 }
@@ -266,7 +272,7 @@ API.initializeStreams = function (
  * If in a browser, this calls setStdinError.
  */
 function setDefaultStdin() {
-  if (IN_NODE) {
+  if (RUNTIME_ENV.IN_NODE) {
     setStdin(new NodeReader(process.stdin.fd));
   } else {
     setStdin({ stdin: () => prompt() });
@@ -428,7 +434,7 @@ function _setStdwrite(
  * If in a browser, sets stdout to write to console.log and sets isatty(stdout) to false.
  */
 function _getStdoutDefaults(): StdwriteOpts & Partial<Writer> {
-  if (IN_NODE) {
+  if (RUNTIME_ENV.IN_NODE) {
     return new NodeWriter(process.stdout.fd);
   } else {
     return { batched: (x) => console.log(x) };
@@ -441,7 +447,7 @@ function _getStdoutDefaults(): StdwriteOpts & Partial<Writer> {
  * If in a browser, sets stdout to write to console.log and sets isatty(stdout) to false.
  */
 function _getStderrDefaults(): StdwriteOpts & Partial<Writer> {
-  if (IN_NODE) {
+  if (RUNTIME_ENV.IN_NODE) {
     return new NodeWriter(process.stderr.fd);
   } else {
     return { batched: (x) => console.warn(x) };
