@@ -1,8 +1,8 @@
-import * as chai from "chai";
-import sinon from "sinon";
+import assert from "node:assert/strict";
+import { afterEach, beforeEach, describe, it } from "node:test";
+import { calculateInstallBaseUrl } from "../../compat";
 import { PackageManager, toStringArray } from "../../load-package.ts";
 import { genMockAPI, genMockModule } from "./test-helper.ts";
-import { calculateInstallBaseUrl } from "../../load-package.ts";
 
 describe("PackageManager", () => {
   it("should initialize with API and Module", () => {
@@ -13,35 +13,32 @@ describe("PackageManager", () => {
 });
 
 describe("logStdout and logStderr", () => {
-  it("Should use console.log and console.error if no logger is provided", () => {
+  it("Should use console.log and console.error if no logger is provided", (t) => {
     const mockApi = genMockAPI();
     const mockMod = genMockModule();
 
     const pm = new PackageManager(mockApi, mockMod);
 
-    const logSpy = sinon.spy(pm, "stdout");
-    const errorSpy = sinon.spy(pm, "stderr");
+    const logSpy = t.mock.method(pm, "stdout");
+    const errorSpy = t.mock.method(pm, "stderr");
 
     pm.logStdout("stdout");
     pm.logStderr("stderr");
 
-    chai.assert.isTrue(logSpy.calledOnce);
-    chai.assert.isTrue(errorSpy.calledOnce);
-    chai.assert.isTrue(logSpy.calledWith("stdout"));
-    chai.assert.isTrue(errorSpy.calledWith("stderr"));
-
-    logSpy.restore();
-    errorSpy.restore();
+    assert.equal(logSpy.mock.callCount(), 1);
+    assert.equal(errorSpy.mock.callCount(), 1);
+    assert.equal(logSpy.mock.calls[0].arguments[0], "stdout");
+    assert.equal(errorSpy.mock.calls[0].arguments[0], "stderr");
   });
 
-  it("Should be overwritten when setCallbacks is called", () => {
+  it("Should be overwritten when setCallbacks is called", (t) => {
     const mockApi = genMockAPI();
     const mockMod = genMockModule();
 
     const pm = new PackageManager(mockApi, mockMod);
 
-    const stdoutLogger = sinon.spy();
-    const stderrLogger = sinon.spy();
+    const stdoutLogger = t.mock.method(pm, "stdout");
+    const stderrLogger = t.mock.method(pm, "stderr");
 
     pm.setCallbacks(
       stdoutLogger,
@@ -51,23 +48,20 @@ describe("logStdout and logStderr", () => {
       pm.logStderr("stderr");
     })();
 
-    chai.assert.isTrue(stdoutLogger.calledOnce);
-    chai.assert.isTrue(stderrLogger.calledOnce);
-    chai.assert.isTrue(stdoutLogger.calledWith("stdout"));
-    chai.assert.isTrue(stderrLogger.calledWith("stderr"));
+    assert.equal(stdoutLogger.mock.callCount(), 1);
+    assert.equal(stderrLogger.mock.callCount(), 1);
+    assert.equal(stdoutLogger.mock.calls[0].arguments[0], "stdout");
+    assert.equal(stderrLogger.mock.calls[0].arguments[0], "stderr");
   });
 });
 
 describe("toStringArray", () => {
   it("Should convert string to array of strings", () => {
-    chai.assert.deepEqual(toStringArray("hello"), ["hello"]);
+    assert.deepEqual(toStringArray("hello"), ["hello"]);
   });
 
   it("Should return the array if it is already an array", () => {
-    chai.assert.deepEqual(toStringArray(["hello", "world"]), [
-      "hello",
-      "world",
-    ]);
+    assert.deepEqual(toStringArray(["hello", "world"]), ["hello", "world"]);
   });
 
   it("Should convert PyProxy to array of strings", () => {
@@ -76,7 +70,7 @@ describe("toStringArray", () => {
       toJs: () => ["hello", "world"],
     };
 
-    chai.assert.deepEqual(toStringArray(pyProxyMock), ["hello", "world"]);
+    assert.deepEqual(toStringArray(pyProxyMock), ["hello", "world"]);
   });
 });
 
@@ -91,33 +85,33 @@ describe("getLoadedPackageChannel", () => {
     };
 
     const loadedPackage = pm.getLoadedPackageChannel("package");
-    chai.assert.equal(loadedPackage, "channel");
+    assert.equal(loadedPackage, "channel");
 
     const notLoadedPackage = pm.getLoadedPackageChannel("notLoadedPackage");
-    chai.assert.equal(notLoadedPackage, null);
+    assert.equal(notLoadedPackage, null);
   });
 
   describe("streamReady and flushing buffers", () => {
-    it("Should flush stdout and stderr buffers when stream is ready", () => {
+    it("Should flush stdout and stderr buffers when stream is ready", (t) => {
       const mockApi = genMockAPI();
       const mockMod = genMockModule();
 
-      const logStdoutSpy = sinon.spy(mockMod, "_print_stdout");
-      const logStderrSpy = sinon.spy(mockMod, "_print_stderr");
+      const logStdoutSpy = t.mock.method(mockMod, "_print_stdout");
+      const logStderrSpy = t.mock.method(mockMod, "_print_stderr");
 
       const pm = new PackageManager(mockApi, mockMod);
       pm.logStdout("stdout message");
       pm.logStderr("stderr message");
 
       // not called yet, buffers should not be flushed
-      chai.assert.isFalse(logStdoutSpy.called);
-      chai.assert.isFalse(logStderrSpy.called);
+      assert.equal(logStdoutSpy.mock.callCount(), 0);
+      assert.equal(logStderrSpy.mock.callCount(), 0);
 
       pm.flushBuffers();
 
       // now buffers should be flushed
-      chai.assert.isTrue(logStdoutSpy.calledOnce);
-      chai.assert.isTrue(logStderrSpy.calledOnce);
+      assert.equal(logStdoutSpy.mock.callCount(), 1);
+      assert.equal(logStderrSpy.mock.callCount(), 1);
     });
   });
 });
@@ -143,34 +137,34 @@ describe("calculateInstallBaseUrl", () => {
     const result = calculateInstallBaseUrl(
       "https://cdn.example.com/pyodide/pyodide-lock.json",
     );
-    chai.assert.equal(result, "https://cdn.example.com/pyodide/");
+    assert.equal(result, "https://cdn.example.com/pyodide/");
   });
 
   it("Should extract base URL from file URL", () => {
     const result = calculateInstallBaseUrl(
       "file:///tmp/pyodide/pyodide-lock.json",
     );
-    chai.assert.equal(result, "file:///tmp/pyodide/");
+    assert.equal(result, "file:///tmp/pyodide/");
   });
 
   it("Should extract base URL from relative URL with path", () => {
     const result = calculateInstallBaseUrl("./pyodide/pyodide-lock.json");
-    chai.assert.equal(result, "./pyodide/");
+    assert.equal(result, "./pyodide/");
   });
 
   it("Should extract base URL from relative URL with parent directory", () => {
     const result = calculateInstallBaseUrl("../pyodide/pyodide-lock.json");
-    chai.assert.equal(result, "../pyodide/");
+    assert.equal(result, "../pyodide/");
   });
 
   it("Should handle URL with no path component", () => {
     const result = calculateInstallBaseUrl("pyodide-lock.json");
-    chai.assert.equal(result, ".");
+    assert.equal(result, ".");
   });
 
   it("Should handle empty string", () => {
     const result = calculateInstallBaseUrl("");
-    chai.assert.equal(result, ".");
+    assert.equal(result, ".");
   });
 
   it("Should fallback to location when URL has no slash", () => {
@@ -180,7 +174,7 @@ describe("calculateInstallBaseUrl", () => {
     };
 
     const result = calculateInstallBaseUrl("pyodide-lock.json");
-    chai.assert.equal(result, "https://example.com/app/");
+    assert.equal(result, "https://example.com/app/");
   });
 
   it("Should fallback to location when URL is empty", () => {
@@ -190,7 +184,7 @@ describe("calculateInstallBaseUrl", () => {
     };
 
     const result = calculateInstallBaseUrl("");
-    chai.assert.equal(result, "https://example.com/app/");
+    assert.equal(result, "https://example.com/app/");
   });
 
   it("Should fallback to '.' when no location available", () => {
@@ -198,34 +192,34 @@ describe("calculateInstallBaseUrl", () => {
     delete (globalThis as any).location;
 
     const result = calculateInstallBaseUrl("pyodide-lock.json");
-    chai.assert.equal(result, ".");
+    assert.equal(result, ".");
   });
 
   it("Should handle URL with query parameters", () => {
     const result = calculateInstallBaseUrl(
       "https://cdn.example.com/pyodide/pyodide-lock.json?v=1.0",
     );
-    chai.assert.equal(result, "https://cdn.example.com/pyodide/");
+    assert.equal(result, "https://cdn.example.com/pyodide/");
   });
 
   it("Should handle URL with hash fragment", () => {
     const result = calculateInstallBaseUrl(
       "https://cdn.example.com/pyodide/pyodide-lock.json#section",
     );
-    chai.assert.equal(result, "https://cdn.example.com/pyodide/");
+    assert.equal(result, "https://cdn.example.com/pyodide/");
   });
 
   it("Should handle URL with both query parameters and hash", () => {
     const result = calculateInstallBaseUrl(
       "https://cdn.example.com/pyodide/pyodide-lock.json?v=1.0#section",
     );
-    chai.assert.equal(result, "https://cdn.example.com/pyodide/");
+    assert.equal(result, "https://cdn.example.com/pyodide/");
   });
 
   it("Should handle URL with username and password", () => {
     const result = calculateInstallBaseUrl(
       "https://user:pass@cdn.example.com/pyodide/pyodide-lock.json",
     );
-    chai.assert.equal(result, "https://user:pass@cdn.example.com/pyodide/");
+    assert.equal(result, "https://user:pass@cdn.example.com/pyodide/");
   });
 });
