@@ -1,7 +1,8 @@
+import re
 import sys
 from collections.abc import Callable, Sequence
 from importlib.abc import Loader, MetaPathFinder
-from importlib.machinery import ModuleSpec
+from importlib.machinery import ModuleSpec, PathFinder
 from importlib.util import spec_from_loader
 from types import ModuleType
 from typing import Any
@@ -116,6 +117,44 @@ class JsLoader(Loader):
         return True
 
 
+WINDOWS_DRIVE_REGEX = re.compile(r"^[a-zA-Z]:")
+
+
+class WindowsToLinuxPathFinder:
+    """
+    A MetaPathFinder that converts Windows-style paths in sys.path to
+    Linux-style paths for module searching.
+
+    This is useful when running Pyodide CLI in a Windows environment where
+    users or system configurations might add Windows-style paths to sys.path.
+    """
+
+    @classmethod
+    def find_spec(
+        cls,
+        fullname: str,
+        path: Sequence[str] | None = None,
+        target: ModuleType | None = None,
+    ) -> ModuleSpec | None:
+        sys_path = path if path is not None else sys.path
+
+        converted_paths: list[str] = []
+        for p in sys_path:
+            # Convert only if it's a string and looks like a Windows absolute path
+            if not WINDOWS_DRIVE_REGEX.match(p):
+                continue
+
+            # Convert 'C:\path\to\dir' to '/path/to/dir'
+            # Sometimes the path may contain forward slashes, sometimes backslashes...
+            linux_path = p[2:].replace("//", "/").replace("\\", "/")
+            converted_paths.append(linux_path)
+
+        if not converted_paths:
+            return None
+
+        return PathFinder.find_spec(fullname, converted_paths, target)
+
+
 jsfinder: JsFinder = JsFinder()
 register_js_module = jsfinder.register_js_module
 unregister_js_module = jsfinder.unregister_js_module
@@ -137,6 +176,21 @@ def register_js_finder() -> None:
         if isinstance(importer, JsFinder):
             raise RuntimeError("JsFinder already registered")
     sys.meta_path.append(jsfinder)
+
+
+def register_windows_finder() -> None:
+    """A bootstrap function to register WindowsToLinuxPathFinder in sys.meta_path.
+
+    This is called in `loadPyodide` in `pyodide.js` to allow Windows-style paths
+    in sys.path to be converted to Linux-style paths for module searching.
+
+    Using class instead of instance to alleviate the need for instantiation
+    (https://docs.python.org/3/library/importlib.html#importlib.machinery.PathFinder)
+    """
+    for importer in sys.meta_path:
+        if importer is WindowsToLinuxPathFinder:
+            raise RuntimeError("WindowsToLinuxPathFinder already registered")
+    sys.meta_path.append(WindowsToLinuxPathFinder)
 
 
 STDLIBS = sys.stdlib_module_names | {"test"}
