@@ -43,11 +43,24 @@ class _JsObject(metaclass=_JsMeta):
 
 class XMLHttpRequest(_JsObject):
     response: str
+    responseText: str
+    responseURL: str
+    status: int
+    statusText: str
 
     @staticmethod
     def new() -> XMLHttpRequest: ...
-    def open(self, method: str, url: str, sync: bool) -> None: ...
-    def send(self, body: JsProxy | None = None) -> None: ...
+    def open(
+        self,
+        method: str,
+        url: str,
+        sync: bool,
+        username: str | None = None,
+        password: str | None = None,
+    ) -> None: ...
+    def send(self, body: str | bytes | JsProxy | None = None) -> None: ...
+    def setRequestHeader(self, name: str, value: str) -> None: ...
+    def getAllResponseHeaders(self) -> str: ...
 
 class Object(_JsObject):
     @staticmethod
@@ -76,6 +89,9 @@ class Uint8Array(_TypedArray):
     BYTES_PER_ELEMENT = 1
 
 class Float64Array(_TypedArray):
+    BYTES_PER_ELEMENT = 8
+
+class BigInt64Array(_TypedArray):
     BYTES_PER_ELEMENT = 8
 
 class JSON(_JsObject):
@@ -147,3 +163,13 @@ class Response(_JsObject):
 class Promise(_JsObject):
     @staticmethod
     def resolve(value: Any) -> Promise: ...
+
+class Request(_JsObject):
+    @overload
+    @staticmethod
+    def new(url: str, options: Any) -> Request: ...
+    @overload
+    @staticmethod
+    def new(url: str, **kwargs: Any) -> Request: ...
+    @property
+    def url(self) -> str: ...
