@@ -8,9 +8,9 @@
 # the _pyodide package.
 #
 # This package is imported by the test suite as well, and currently we don't use
-# pytest mocks for js or pyodide_js, so make sure to test "if IN_BROWSER" before
+# pytest mocks for js or pyodide_js, so make sure to test "if IN_PYODIDE" before
 # importing from these.
-__version__ = "0.29.0.dev0"
+__version__ = "0.30.0.dev0"
 
 __all__ = ["__version__", "console", "code", "ffi", "http", "webloop"]
 
